+++ src/accessories/AccessoryService.ts
@@ -15,6 +15,7 @@ import { OccupancySensor } from './OccupancySensor';
 import { GarageDoorOpener } from './GarageDoorOpener';
 import { UnitSwitch } from './UnitSwitch';
 import { UnitLightbulb } from './UnitLightbulb';
+import { Thermostat } from './Thermostat';
 
 export class AccessoryService {
   private accessories: Map<string, AccessoryBase> = new Map();
@@ -39,6 +40,7 @@ export class AccessoryService {
       this.discoverGarageDoors();
       this.discoverUnitSwitches();
       this.discoverUnitLightbulbs();
+      this.discoverThermostats();
     } catch (error) {
       this.platform.log.error(error);
     }
@@ -406,6 +408,30 @@ export class AccessoryService {
     }
   }
 
+  discoverThermostats(): void {
+    this.platform.log.debug(this.constructor.name, 'discoverThermostats');
+
+    const thermostats = new Map<number, string>();
+
+    if (this.platform.settings.includeThermostats) {
+      for(const [index, thermostat] of this.platform.omniService.thermostats) {
+        thermostats.set(index, thermostat.name);
+      }
+    }
+
+    for(const [index, name] of thermostats) {
+      this.addAccessory(Thermostat, Thermostat.type, name, index);
+    }
+
+    for(const accessory of this.accessories.values()) {
+      if (accessory instanceof Thermostat) {
+        if (!thermostats.has(accessory.platformAccessory.context.index)) {
+          this.removeAccessory(Thermostat.type, accessory.platformAccessory.context.index);
+        }
+      }
+    }
+  }
+
   addAccessory<TAccessory extends AccessoryBase>(
     Accessory: new (platform: OmniLinkPlatform, accessory: PlatformAccessory) => TAccessory,
     type: string, name: string, index?: number,
@@ -491,6 +517,8 @@ export class AccessoryService {
         return new UnitSwitch(this.platform, platformAccessory);
       case 'unitlightbulb':
         return new UnitLightbulb(this.platform, platformAccessory);
+      case 'thermostat':
+        return new Thermostat(this.platform, platformAccessory);
     }
   }